Implement the q_ascend() function
The `last` pointer will track the node being considered for removal,
while the `front` pointer will traverse the queue. If the node value
pointed to by the `front` pointer is smaller than the node value
pointed to by the `last` pointer, then the node pointed to by the
`last` pointer will be removed. These actions will repeat until the
`last` pointer points to the tail node.