Re-implement q_delete_dup()
In the previous version, not all duplicate strings could be deleted.
To address this, a method similar to `descend()` was adapted:
the `last` pointer indicates the string to be evaluated for duplication.
If the value pointed to by the `front` pointer matches that of the
`last` pointer, we remove the `front` pointer until it reaches
the head pointer. Finally, we delete the `last` pointer.